# Document/swager/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a sample server Userinformation server.  You can find \nout\
    \ more about Swagger at \n"
  version: "1.0.0"
  title: "User Information"
  contact:
    email: "spraghuwanshi@gmail.com"
host: "virtserver.swaggerhub.com"
basePath: "/sunmoon/test/1.0.0"
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "callId"
        in: "header"
        description: "Unique identifier from the service consumer. E.g. Reference\
          \ number while creating a userinf0/Address from UI can be passed through\
          \ all APIs to provide visibility across calls."
        required: true
        type: "string"
      - name: "channel"
        in: "header"
        description: "Name of the channel - B2B/CES"
        required: true
        type: "string"
      - name: "x-mytoll-identity"
        in: "header"
        description: "UserID (Principal)"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      operationId: "loginUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "callId"
        in: "header"
        description: "Unique identifier from the service consumer. E.g. Reference\
          \ number while creating a userinf0/Address from UI can be passed through\
          \ all APIs to provide visibility across calls."
        required: true
        type: "string"
      - name: "channel"
        in: "header"
        description: "Name of the channel - B2B/CES"
        required: true
        type: "string"
      - name: "x-mytoll-identity"
        in: "header"
        description: "UserID (Principal)"
        required: true
        type: "string"
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid username/password supplied"
        500:
          description: "Error Occured."
  /user/logout:
    get:
      tags:
      - "user"
      summary: "Logs out current logged in user session"
      operationId: "logoutUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "callId"
        in: "header"
        description: "Unique identifier from the service consumer. E.g. Reference\
          \ number while creating a userinf0/Address from UI can be passed through\
          \ all APIs to provide visibility across calls."
        required: true
        type: "string"
      - name: "channel"
        in: "header"
        description: "Name of the channel - B2B/CES"
        required: true
        type: "string"
      - name: "x-mytoll-identity"
        in: "header"
        description: "UserID (Principal)"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
  /user/{username}:
    get:
      tags:
      - "user"
      summary: "Get user by user name"
      operationId: "getUserByName"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "callId"
        in: "header"
        description: "Unique identifier from the service consumer. E.g. Reference\
          \ number while creating a userinf0/Address from UI can be passed through\
          \ all APIs to provide visibility across calls."
        required: true
        type: "string"
      - name: "channel"
        in: "header"
        description: "Name of the channel - B2B/CES"
        required: true
        type: "string"
      - name: "x-mytoll-identity"
        in: "header"
        description: "UserID (Principal)"
        required: true
        type: "string"
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "callId"
        in: "header"
        description: "Unique identifier from the service consumer. E.g. Reference\
          \ number while creating a userinf0/Address from UI can be passed through\
          \ all APIs to provide visibility across calls."
        required: true
        type: "string"
      - name: "channel"
        in: "header"
        description: "Name of the channel - B2B/CES"
        required: true
        type: "string"
      - name: "x-mytoll-identity"
        in: "header"
        description: "UserID (Principal)"
        required: true
        type: "string"
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
securityDefinitions:
  apiKey:
    type: "apiKey"
    name: "apiKey"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      AddressParty1:
        $ref: "#/definitions/AddressParty1"
      AddressParty2:
        $ref: "#/definitions/AddressParty2"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"
  AddressParty1:
    type: "object"
    properties:
      AddressId:
        type: "integer"
      CompanyName:
        type: "string"
      PartyName:
        type: "string"
      WorkPhoneNumber:
        type: "string"
      MobileNumber:
        type: "string"
      Email:
        type: "string"
      PostCode:
        type: "string"
      State:
        type: "string"
      CountryCode:
        type: "string"
      AddressNote:
        type: "string"
      AccountCode:
        type: "string"
      ReceiverAddressType:
        type: "string"
      BusinesstaxNumber:
        type: "string"
      AddressType:
        type: "string"
      CreatedTs:
        type: "string"
        format: "date-time"
      CreatedBy:
        type: "string"
      LastModifiedTs:
        type: "string"
        format: "date-time"
      LastModifiedBy:
        type: "string"
    xml:
      name: "AddressParty1"
  AddressParty2:
    type: "object"
    properties:
      AddressId:
        type: "integer"
      CompanyName:
        type: "string"
      PartyName:
        type: "string"
      WorkPhoneNumber:
        type: "string"
      MobileNumber:
        type: "string"
      Email:
        type: "string"
      PostCode:
        type: "string"
      State:
        type: "string"
      CountryCode:
        type: "string"
      AddressNote:
        type: "string"
      AccountCode:
        type: "string"
      ReceiverAddressType:
        type: "string"
      BusinesstaxNumber:
        type: "string"
      AddressType:
        type: "string"
      CreatedTs:
        type: "string"
        format: "date-time"
      CreatedBy:
        type: "string"
      LastModifiedTs:
        type: "string"
        format: "date-time"
      LastModifiedBy:
        type: "string"
    xml:
      name: "AddressParty2"
parameters:
  CallerIdHeader:
    name: "callId"
    in: "header"
    description: "Unique identifier from the service consumer. E.g. Reference number\
      \ while creating a userinf0/Address from UI can be passed through all APIs to\
      \ provide visibility across calls."
    required: true
    type: "string"
  ChannelHeader:
    name: "channel"
    in: "header"
    description: "Name of the channel - B2B/CES"
    required: true
    type: "string"
  UserPrincipalHeader:
    name: "x-mytoll-identity"
    in: "header"
    description: "UserID (Principal)"
    required: true
    type: "string"